LayoutTests/css3/viewport-percentage-lengths/viewport-percentage-lengths-resize.html

<!DOCTYPE html>
<style>
#test {
    font-size: 5vh;
    width: 50vw;
}
#testpseudo:after {
    margin-left: 20vmin;
    padding-right: 25vmax;
    content: '';
}
</style>
<body>
    <div id="test"></div>
    <div id="testpseudo"></div>
</body>
<script src="../../resources/js-test-pre.js"></script>
<script src="resources/resize-test.js"></script>
<script>
description("Test that viewport lengths and font sizes properly resize with the viewport.");
standardResizeTest(function (size) {
        var min = Math.min(window.innerWidth, window.innerHeight);
        var max = Math.max(window.innerWidth, window.innerHeight);
        shouldBe("getComputedStyle(test).fontSize", "'" + window.innerHeight/20 + "px'");
        shouldBe("getComputedStyle(test).width", "'" + window.innerWidth/2 + "px'");
        shouldBe("getComputedStyle(testpseudo, ':after').marginLeft", "'" + min/5 + "px'");
        shouldBe("getComputedStyle(testpseudo, ':after').paddingRight", "'" + max/4 + "px'");
});
</script>
<script src="../../resources/js-test-post.js"></script>
